Add CLI support for CycloneDX interfaces.
